Fixing API rate limiting
The reddit API was rate limiting to 1 request every 3 seconds. Since the
script didn't account for this it would die after 1 request.

I added a simple time.sleep(4) to this to avoid being rate limited.

Since 'time' was already a variable that was being used, I refactored that
variable to be 'created_time'.